refator: 간편 회원가입 간편 로그인 수정 #57
