+++ app/(home)/layout.tsx
@@ -1,23 +1,9 @@
-'use client';
-
-import Header from '@/components/common/Header';
-import { useAuthStore } from '@/providers/auth';
-import { useSession } from 'next-auth/react';
-import { useEffect } from 'react';
+import HeaderWrapper from './header-wrapper';
 
 function HomeLayout({ children }: { children: React.ReactNode | JSX.Element }) {
-  const { data: session } = useSession();
-  const oAuthLogin = useAuthStore((state) => state.oAuthLogin);
-  useEffect(() => {
-    if (session?.provider?.toUpperCase() === 'GOOGLE') {
-      if (session.id_token) {
-        oAuthLogin('GOOGLE', { token: session.id_token });
-      }
-    }
-  }, [oAuthLogin, session?.provider, session?.id_token]);
   return (
     <>
-      <Header />
+      <HeaderWrapper />
       {children}
     </>
   );
